restructures tasks/scripts
